Refactor utils into discord_utils and game_utils
