+++ .gitignore
@@ -4,4 +4,4 @@ main.db
 config/app_config.json
 .cache/
 logs/
-built/
+build/
