+++ src/commands/ransom.ts
@@ -1,5 +1,5 @@
 import BaseCommand, { CommandArgs } from "./base_command";
-import { EMBED_INFO_COLOR } from "../helpers/utils";
+import { EMBED_INFO_COLOR } from "../helpers/discord_utils";
 
 class RansomCommand implements BaseCommand {
     call({ message }: CommandArgs) {
