+++ src/commands/volume.ts
@@ -1,5 +1,6 @@
 import BaseCommand, { CommandArgs } from "./base_command";
-import { sendOptionsMessage, getDebugContext, GameOptions } from "../helpers/utils";
+import { sendOptionsMessage, getDebugContext } from "../helpers/discord_utils";
+import { GameOptions } from "../helpers/game_utils";
 import _logger from "../logger";
 const logger = _logger("volume");
 const DEFAULT_VOLUME = 50;
